plotting - Add ViewManager class to update basemap whenever axis limits change
This is a very basic implementation that fetches tiles in response to changing axis limits.
It is helpful for data exploration, but it does block the main thread.
It requires a cache to be effective. (feature/cache-tiles)

An improved implementation (more responsive) would have to expose part of bounds2img to update each
tile as it is retrieved. Bearing in mind that matplotlib cannot redraw except in the main thread, this would be tricky.