Add PDOK providers.
Note: not a 100% of them work. The issue appears to be some inconsistency in the URL (which does not match PDOK's own description...). This appears to be rather common case with these WMTS services. In this specific case, however, owslib IS able to fetch a tile, so there's like something that can be done.

There are also cases where the WMTS URL structure looks nothing like what owslib expects. I suppose manual tweaking is required in that case.